Update some links
I couldn't find a new link for Swofford et al's Phylogenetic inference in Molecular Systematics from Sauer, which is now OUP - the existing Sauer link currently redirects to the generic OUP homepage.

The R-ape link also needs updating, but that auto-updated when I ran bookdown so it's not included here.